fix(lint): split prompt-plan resolution out of launch-agent-in-new-tab
launch-agent-in-new-tab.ts sat at exactly 300 counted lines; the oxfmt
line-width wrapping in the live-port indicator merge pushed it to 303,
failing max-lines for every PR. Per repo policy (no max-lines bumps),
extract the prompt-delivery plan branching into
agent-launch-prompt-plan.ts as a pure function; no behavior change.
